refactor configuration to use env variables.
Creating utilities when configuring stuff with ZCML did not work well in
tests, because the utility was not available in the other thread when
doing an actual request with testbrowser.
I had this problem in different situations and could not solve
it. Therefore I refactored everything to using environment variables.